+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Expect active development and potentially significant breaking changes in the `0
 ### vNEXT
 
 - ...
+
+### 0.5.22
 - Fix bug that caused updateQuery and reducers to run on stopped queries [PR #1054](https://github.com/apollostack/apollo-client/pull/1054)
 - Ensure transporters are using `isomorphic-fetch` instead of `whatwg-fetch` for universal compatibility [PR #1018](https://github.com/apollostack/apollo-client/pull/1018)